pyproject.toml
[project]
name = "dataset-augmenter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "albumentationsx>=2.2.6",
    "dotenv>=0.9.9",
    "fastapi>=0.136.1",
    "loguru>=0.7.3",
    "matplotlib>=3.10.9",
    "numpy>=2.4.4",
    "opencv-python-headless>=4.13.0.92",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.3",
    "tqdm>=4.67.3",
    "uvicorn>=0.46.0",
]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
